In DataRepeater, override the controllerFindAndInstance callback instead of the _controller_changed observer.
The purpose of this code is to give us a hook to refresh the DataRepeater after its controller changes, but we were refreshing too early (in some cases before the new controller had been found). If we refresh in the callback, we know we have our new controller.

Enyo-DCO-1.1-Signed-Off-By: Gray Norton <gray.norton@lge.com>